move the import_all on the views to app.py
previously, the import_all on the views was done in the views
__init__.py, but this caused a circualar import error when trying to
add something like "from tahrir.app import csrf" to a view (i.e. root.py)

moving the import_all to just before registrering the blueprints is how
it is done in noggin -- so just copied that pattern in this commit.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>